+++ R/find_parameters_mixed.R
@@ -154,6 +154,33 @@ find_parameters.nlmerMod <- function(x,
 }
 
 
+#' @export
+find_parameters.hglm <- function(x,
+                                 effects = c("all", "fixed", "random"),
+                                 component = c("all", "conditional", "dispersion"),
+                                 flatten = FALSE,
+                                 ...) {
+  effects <- match.arg(effects)
+  fe <- x$fixef
+  re <- x$ranef
+
+  f <- find_formula(x)
+  if (!is.null(f$dispersion)) {
+    disp <- summary(x)$SummVC1
+    disp_name <- rownames(disp)
+  } else {
+    disp_name <- NULL
+  }
+
+  l <- compact_list(list(
+    conditional = names(fe),
+    random = names(re),
+    dispersion = disp_name
+  ))
+
+  .filter_parameters(l, effects = effects, component = component, flatten = flatten)
+}
+
 
 #' @rdname find_parameters.glmmTMB
 #' @export