+++ R/get_data.R
@@ -101,6 +101,13 @@ get_data <- function(x, ...) {
     {
       # recover data frame from environment
       dat <- .recover_data_from_environment(x, data_name = data_name)
+      # for metafor, we need to add weights...
+      if (inherits(x, c("rma.uni", "rma"))) {
+        ## TODO: check if we need to do this for other meta-analysis packages, too
+        wdat <- data.frame(Weights = get_weights(x))
+        additional_variables <- c(additional_variables, "Weights")
+        dat <- tryCatch(cbind(dat, wdat), error = function(e) dat)
+      }
       # additional variables? Some models, like plm::plm(), have an "index"
       # slot in the model call with further variables
       if (!is.null(additional_variables) && !isTRUE(additional_variables)) {
@@ -202,22 +209,16 @@ get_data <- function(x, ...) {
   }
 
   # first, try environment of formula, see #666
-  dat <- tryCatch(eval(model_call$data, envir = environment(model_call$formula)),
-    error = function(e) NULL
-  )
+  dat <- .safe(eval(model_call$data, envir = environment(model_call$formula)))
 
   # second, try to extract formula directly
   if (is.null(dat)) {
-    dat <- tryCatch(eval(model_call$data, envir = environment(find_formula(x))),
-      error = function(e) NULL
-    )
+    dat <- .safe(eval(model_call$data, envir = environment(find_formula(x))))
   }
 
   # next try, parent frame
   if (is.null(dat)) {
-    dat <- tryCatch(eval(model_call$data, envir = parent.frame()),
-      error = function(e) NULL
-    )
+    dat <- .safe(eval(model_call$data, envir = parent.frame()))
   }
 
   # sanity check- if data frame is named like a function, e.g.
@@ -225,21 +226,17 @@ get_data <- function(x, ...) {
   # we then need to reset "dat" to NULL and search in the global env
 
   if (!is.null(dat) && !is.data.frame(dat)) {
-    dat <- tryCatch(as.data.frame(dat), error = function(e) NULL)
+    dat <- .safe(as.data.frame(dat))
   }
 
   # third try, global env
   if (is.null(dat)) {
-    dat <- tryCatch(eval(model_call$data, envir = globalenv()),
-      error = function(e) NULL
-    )
+    dat <- .safe(eval(model_call$data, envir = globalenv()))
   }
 
   # last try, internal env
   if (is.null(dat)) {
-    dat <- tryCatch(eval(model_call$data, envir = parent.env(x$call_env)),
-      error = function(e) NULL
-    )
+    dat <- .safe(eval(model_call$data, envir = parent.env(x$call_env)))
   }
 
   if (!is.null(dat) && object_has_names(model_call, "subset")) {
@@ -704,16 +701,13 @@ get_data.merMod <- function(x,
   # fall back to extract data from model frame
   effects <- match.arg(effects, choices = c("all", "fixed", "random"))
 
-  mf <- tryCatch(
+  mf <- .safe(
     {
       switch(effects,
         fixed = stats::model.frame(x, fixed.only = TRUE),
         all = stats::model.frame(x, fixed.only = FALSE),
         random = stats::model.frame(x, fixed.only = FALSE)[, find_random(x, split_nested = TRUE, flatten = TRUE), drop = FALSE]
       )
-    },
-    error = function(x) {
-      NULL
     }
   )
   .prepare_get_data(x, mf, effects, verbose = verbose)
@@ -755,7 +749,7 @@ get_data.mmrm_tmb <- get_data.mmrm
 
 #' @export
 get_data.merModList <- function(x, effects = "all", ...) {
-  warning("Can't access data for `merModList` objects.", call. = FALSE)
+  format_warning("Can't access data for `merModList` objects.")
   return(NULL)
 }
 
@@ -776,16 +770,13 @@ get_data.MANOVA <- function(x,
   # fall back to extract data from model frame
   effects <- match.arg(effects, choices = c("all", "fixed", "random"))
 
-  mf <- tryCatch(
+  mf <- .safe(
     {
       switch(effects,
         fixed = .remove_column(x$input$data, x$input$subject),
         all = x$input$data,
         random = x$input$data[, x$input$subject, drop = FALSE]
       )
-    },
-    error = function(x) {
-      NULL
     }
   )
   .prepare_get_data(x, mf, effects, verbose = verbose)
@@ -853,7 +844,7 @@ get_data.glmm <- function(x,
   effects <- match.arg(effects, choices = c("all", "fixed", "random"))
   dat <- get_data.default(x, verbose = verbose)
 
-  mf <- tryCatch(
+  mf <- .safe(
     {
       switch(effects,
         fixed = dat[, find_predictors(
@@ -865,9 +856,6 @@ get_data.glmm <- function(x,
         all = dat,
         random = dat[, find_random(x, split_nested = TRUE, flatten = TRUE), drop = FALSE]
       )
-    },
-    error = function(x) {
-      NULL
     }
   )
   .prepare_get_data(x, mf, effects, verbose = verbose)
@@ -939,16 +927,13 @@ get_data.glmmadmb <- function(x,
   fixed_data <- x$frame
   random_data <- .recover_data_from_environment(x)[, find_random(x, split_nested = TRUE, flatten = TRUE), drop = FALSE]
 
-  mf <- tryCatch(
+  mf <- .safe(
     {
       switch(effects,
         fixed = fixed_data,
         all = cbind(fixed_data, random_data),
         random = random_data
       )
-    },
-    error = function(x) {
-      NULL
     }
   )
   .prepare_get_data(x, mf, effects, verbose = verbose)
@@ -1060,7 +1045,7 @@ get_data.lme <- function(x, effects = "all", source = "environment", verbose = T
 
   # fall back to extract data from model frame
   effects <- match.arg(effects, choices = c("all", "fixed", "random"))
-  dat <- tryCatch(x$data, error = function(x) NULL)
+  dat <- .safe(x$data)
 
   stats::na.omit(.get_data_from_modelframe(x, dat, effects))
 }
@@ -1485,7 +1470,7 @@ get_data.ivreg <- function(x, source = "environment", verbose = TRUE, ...) {
   }
 
   # fall back to extract data from model frame
-  mf <- tryCatch(stats::model.frame(x), error = function(e) NULL)
+  mf <- .safe(stats::model.frame(x))
   ft <- find_variables(x, flatten = TRUE)
 
   if (!insight::is_empty_object(mf)) {
@@ -1494,24 +1479,18 @@ get_data.ivreg <- function(x, source = "environment", verbose = TRUE, ...) {
     if (is_empty_object(remain)) {
       final_mf <- mf
     } else {
-      final_mf <- tryCatch(
+      final_mf <- .safe(
         {
           dat <- .recover_data_from_environment(x)
           cbind(mf, dat[, remain, drop = FALSE])
-        },
-        error = function(x) {
-          NULL
         }
       )
     }
   } else {
-    final_mf <- tryCatch(
+    final_mf <- .safe(
       {
         dat <- .recover_data_from_environment(x)
         dat[, ft, drop = FALSE]
-      },
-      error = function(x) {
-        NULL
       }
     )
   }
@@ -2110,9 +2089,39 @@ get_data.mlogit <- function(x, source = "environment", verbose = TRUE, ...) {
 
 
 #' @export
-get_data.rma <- function(x, source = "environment", verbose = TRUE, ...) {
+#' @rdname get_data
+#' @param include_interval For meta-analysis models, should normal-approximation
+#'   confidence intervals be added for each response effect size?
+#' @param transf For meta-analysis models, if intervals are included, a function
+#'   applied to each response effect size and its interval.
+#' @param transf_args For meta-analysis models, an optional list of arguments
+#'   passed to the `transf` function.
+#' @param ci For meta-analysis models, the Confidence Interval (CI) level if
+#'   `include_interval = TRUE`. Default to 0.95 (95%).
+get_data.rma <- function(x,
+                         source = "environment",
+                         verbose = TRUE,
+                         include_interval = FALSE,
+                         transf = NULL,
+                         transf_args = NULL,
+                         ci = 0.95,
+                         ...) {
+  # standard errors and moderators are not found by find_predictors(),
+  # so we need them as additional variables
+  model_call <- get_call(x)
+  additional_variables <- c(
+    safe_deparse(model_call$vi),
+    safe_deparse(model_call$sei),
+    safe_deparse(model_call$mods)
+  )
   # try to recover data from environment
-  model_data <- .get_data_from_environment(x, source = source, verbose = verbose, ...)
+  model_data <- .get_data_from_environment(
+    x,
+    source = source,
+    additional_variables = additional_variables,
+    verbose = verbose,
+    ...
+  )
 
   if (!is.null(model_data)) {
     return(model_data)
@@ -2120,11 +2129,51 @@ get_data.rma <- function(x, source = "environment", verbose = TRUE, ...) {
 
   # fall back to extract data from model frame
   mf <- tryCatch(.recover_data_from_environment(x), error = function(x) NULL)
-  .prepare_get_data(x, stats::na.omit(mf), verbose = verbose)
+  mf_attr <- attributes(mf)
+  mf <- merge(mf, data.frame(Weights = get_weights(x)), by = "row.names", all = TRUE, sort = FALSE)
+  rownames(mf) <- mf$Row.names
+  mf$Row.names <- NULL
+  mostattributes(mf) <- c(
+    attributes(mf)[c("names", "row.names")],
+    mf_attr[c("yi.names", "vi.names", "digits", "class")]
+  )
+  if (isTRUE(include_interval)) {
+    model_response <- tryCatch(mf[[find_response(x)]], error = function(x) NULL)
+    sei <- tryCatch(mf[[model_call$sei]], error = function(x) NULL)
+    if (is.null(sei)) {
+      sei <- tryCatch(sqrt(mf[[model_call$vi]]), error = function(x) NULL)
+    }
+    if (is.null(sei)) {
+      format_error("Could not find `sei` or `vi` for this model.")
+    }
+    mf$ci <- ci
+    mf$CI_low <- model_response - stats::qnorm((1 - ci) / 2, lower.tail = FALSE) * sei
+    mf$CI_high <- model_response + stats::qnorm((1 - ci) / 2, lower.tail = FALSE) * sei
+    if (!is.null(transf)) {
+      if (!is.function(transf)) {
+        format_error("`transf` must be a function.")
+      }
+      if (!is.null(transf_args)) {
+        mf[[find_response(x)]] <- sapply(mf[[find_response(x)]], transf, transf_args)
+        mf$CI_low <- sapply(mf$CI_low, transf, transf_args)
+        mf$CI_high <- sapply(mf$CI_high, transf, transf_args)
+      } else {
+        mf[[find_response(x)]] <- sapply(mf[[find_response(x)]], transf)
+        mf$CI_low <- sapply(mf$CI_low, transf)
+        mf$CI_high <- sapply(mf$CI_high, transf)
+      }
+    }
+  }
+  original_rownames <- rownames(x$X)
+  if (is.null(original_rownames)) {
+    original_rownames <- seq_len(nrow(mf))
+  }
+  .prepare_get_data(x, mf[original_rownames, , drop = FALSE], verbose = verbose)
 }
 
 
 #' @export
+# TODO: Check these
 get_data.metaplus <- get_data.rma
 
 
@@ -2159,10 +2208,12 @@ get_data.meta_bma <- function(x, source = "environment", verbose = TRUE, ...) {
 
 
 #' @export
+# TODO: Check these
 get_data.meta_fixed <- get_data.meta_random
 
 
 #' @export
+# TODO: Check these
 get_data.ivFixed <- get_data.rma
 
 
@@ -2174,13 +2225,10 @@ get_data.bfsl <- function(x, ...) {
 
 #' @export
 get_data.mipo <- function(x, ...) {
-  tryCatch(
+  .safe(
     {
       models <- eval(x$call$object)
       get_data(models$analyses[[1]], ...)
-    },
-    error = function(e) {
-      NULL
     }
   )
 }